requirements.txt
flask==3.1.0
pandas==2.3.2
pdfplumber==0.11.6
scikit-learn==1.6.1
lightgbm==4.6.0
shap==0.46.0
networkx==3.6
reportlab==4.4.0
matplotlib==3.10.1
seaborn==0.13.2
apscheduler==3.11.0
python-louvain==0.16
pydantic>=2.0
confluent-kafka>=2.3.0

# Firebase Auth
firebase-admin>=6.0.0

# Optional: LLM + OCR support
google-genai>=1.0.0
pytesseract>=0.3.10
Pillow>=10.0.0
pdf2image>=1.16.0
